+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   Expired credentials will try to be renewed before triggering a log out (#112)
+-   Support for testtuples which parent is an aggregatetuple (#113)
 
 ### Fixed